Initial Release
Minimal Industry Calculor Web service is now built

+++ aba-industry/industry-data/src/main/java/com/aba/industry/model/fuzzysteve/BlueprintDetails.java
@@ -1,19 +1,16 @@
 /*
  * Copyright 2016 maurerit
  *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
  *
  * http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed
- * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for
- * the specific language governing permissions and limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
 
 package com.aba.industry.model.fuzzysteve;
 
-import com.aba.industry.model.Activity;
+import com.aba.industry.model.IndustryActivities;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
@@ -22,7 +19,7 @@ import java.util.Map;
 @Data
 public class BlueprintDetails {
     @JsonProperty
-    private Integer            maxProductionLimit;
+    private Long               maxProductionLimit;
     @JsonProperty( "productTypeID" )
     private Long               productTypeId;
     @JsonProperty( "productTypeName" )
@@ -30,7 +27,7 @@ public class BlueprintDetails {
     @JsonProperty
     private Integer            productQuantity;
     /**
-     * A map of {@link Activity#activityId} to times in seconds unmodified.
+     * A map of {@link IndustryActivities#activityId} to times in seconds unmodified.
      */
     @JsonProperty( "times" )
     private Map<Integer, Long> timesInSeconds;
